Move DocWithRank to the task plugin
