+++ packages/core/package.json
@@ -29,7 +29,6 @@
   },
   "dependencies": {
     "@anticrm/platform": "~0.6.5",
-    "just-clone": "^3.2.1",
-    "lexorank": "~1.0.4"
+    "just-clone": "^3.2.1"
   }
 }
