+++ plugins/recruit-resources/src/components/CreateApplication.svelte
@@ -16,11 +16,10 @@
   import type { Employee } from '@anticrm/contact'
   import contact from '@anticrm/contact'
   import { Account, Class, Client, Doc, generateId, Ref, SortingOrder } from '@anticrm/core'
-  import { calcRank } from '@anticrm/core'
   import { getResource, OK, Resource, Severity, Status } from '@anticrm/platform'
   import { Card, getClient, UserBox } from '@anticrm/presentation'
   import type { Applicant, Candidate } from '@anticrm/recruit'
-  import type { SpaceWithStates, State } from '@anticrm/task'
+  import { calcRank, SpaceWithStates, State } from '@anticrm/task'
   import task from '@anticrm/task'
   import { Grid, Status as StatusControl } from '@anticrm/ui'
   import { createEventDispatcher } from 'svelte'
